server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "D",
    "UP",
    "YTT",
    "ANN",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "ISC",
    "ICN",
    "INP",
    "PIE",
    "PT",
    "RET",
    "SIM",
    "TID",
    "ARG",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "RUF",
]
ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D107",
    "D203",
    "D212",
    "UP006",
    "UP007",
    "D400",
    "D406",
    "D407",
    "PLC1901",
    "UP035",
    "ANN001",
]
unfixable = ["F401", "F841"]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["ANN001", "ANN002", "ANN003", "D103"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
